+++ clients/src/main/java/org/apache/kafka/common/utils/Timer.java
@@ -158,6 +158,10 @@ public class Timer {
         return Math.max(0, deadlineMs - currentTimeMs);
     }
 
+    public long getDeadlineMs() {
+        return deadlineMs;
+    }
+
     /**
      * Get the current time in milliseconds. This will return the same cached value until the timer
      * has been updated using one of the {@link #update()} methods or {@link #sleep(long)} is used.
