+++ core/src/main/scala/kafka/server/KafkaServer.scala
@@ -323,7 +323,7 @@ class KafkaServer(
         kafkaController = new KafkaController(config, zkClient, time, metrics, brokerInfo, brokerEpoch, tokenManager, brokerFeatures, featureCache, threadNamePrefix)
         kafkaController.startup()
 
-        adminManager = new ZkAdminManager(config, metrics, metadataCache, zkClient)
+        adminManager = new ZkAdminManager(config, metrics, metadataCache, zkClient, enableForwarding)
 
         /* start group coordinator */
         // Hardcode Time.SYSTEM for now as some Streams tests fail otherwise, it would be good to fix the underlying issue
